timer: Explicitly cast ticks to `long int`
`timespec.tv_nsec` is implemented as a `long int` while our
`timer_val_t.ticks` is a `uint32_t`. The value will always
be limited to 999999999 as it's a count of ns, but without
explicit casting this will trigger a narrowing warning.

+++ kernel/arch/dreamcast/include/arch/timer.h
@@ -262,7 +262,7 @@ static inline struct timespec arch_timer_gettime(void) {
     /* Convert from ticks to nanoseconds: each clock tick is 80ns. */
     return (struct timespec){
         .tv_sec = time.secs,
-        .tv_nsec = time.ticks * 80,
+        .tv_nsec = (long int)(time.ticks * 80),
     };
 }
